move njvm to seperate file
